requirements.txt
json-cfg
openai
python-dotenv
